Bugfix, refactor and cleanup code.

+++ app/src/main/java/com/example/destiny/data/models/special/SpecialAttack.java
@@ -3,7 +3,7 @@ package com.example.destiny.data.models.special;
 import com.example.destiny.data.models.adventurer.CombatStatistics;
 
 public abstract class SpecialAttack extends Special{
-    public float attackScaling;
+    public double attackScaling;
     public CombatStatistics combatStats;
     public SpecialAttack(CombatStatistics combatStats, String name) {
         this.combatStats = combatStats;
